refac: Extrapolate an interface for reading Zephyr nvs data.
In many place we are reading the length of the data, allocating memory,
and then using allocated memory to read the actual data. Instead of
repeating it in many places we can use an interface to deal with it and
use when we need to access nvs stored data.

Changelog: Title

Ticket: MEN-7380

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>