feat: do not check if artifact is already installed
The Mender Client used to check if the downloaded artifact was already
installed in the past. However, this check was removed in
https://github.com/mendersoftware/mender/pull/1106/commits/1fb62236abdee6d2ed2452592d289320aea1bfa5
as it prevents the use of the "Force update" option, allowing a device
to take part in the same deployment twice. Today, all the
"already-installed" logic is performed in the backend.

Ticket: MEN-7428
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>