feat: update device`s provides
Update device`s provides by filtering it with clears_provides from the
currently stored data and combining this with the new provides data from
the current deployment. On a reboot, store the new provides data in the
deployment-data.json and write it afterwards - else just write.

Removed the `old way` of storing provides, as that simply appended the
new provides data on each deployment.

Ticket: MEN-7429
Changelog: Title

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>