fix: fixed error code check for boot and flash functions
The documentation explicitly states that the respective `flash_*` and
`boot_*` functions return 0 on success. Hence, we should not rely on all
non-negative return codes being success. Futhermore, a negated errno
code is returned on failure. Thus, we should negate it again before
logging it in an error message to avoid any confusion.

Ticket: MEN-7441
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>