fix: get_user_provided_keys callback is now optional
The `get_user_provided_keys` callback is now optional. Furthermore, if
the callback is set but returns an empty key, the client no longer tries
to fix this by generating or loading keys from the store. Now, keys are
only generated or loaded from the store when the callback is explicitly
set to `NULL`.

Ticket: None
Changelog: Commit
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>