fix: Rewriting scheduler
In Marcin Pasinski words:

Maybe "rewriting" is a big word here. I've been testing several different
approaches. What works as a charm and gets the work done is delayable
work. This is IMO the way to go to implement the final version of the
scheduler.

In shared words:

This commit basically replaces the scheduler with a temporary
alternative implementation as described above. No support for mutex,
semaphors is in place so it can only do one thing in the thread.

Support for Inventory and Add-Ons is also disabled, and some debugging
logging has been kept for future work.

We will rework the scheduler in:
* https://northerntech.atlassian.net/browse/MEN-7536

Changelog: None
Ticket: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>