test: Introduce Unit Tests framework
This commit introduces the structure (and the first test!) for the Unit
Tests for the project.

It utilizes the existing "posix" target to build the MCU client as a
library and then link against the tests.

The chosen framework is GoogleTest. We considered several ones, but we
decided that compiling for C++ can have benefits wrt discovering
pitfalls.

Still few TODOs in the code...

Ticket: MEN-7526

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>