build: Include default cache for POSIX target
This inclusion simplifies user calls to `cmake`. And, in any case, if a
cache value is already present, it won't be overriden so user
preferences are respected.

The unit tests in CI were being linked with the "weak" implementations,
that is why the missing dependencies there went missing. And, once
touching the GitLab CI file, let's tidy it up a bit ;)

Ticket: MEN-7526

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>