fix: Only return MENDER_DONE from update_work_function() when rebooting
Returning `MENDER_DONE` from
`mender_client_update_work_function()` signals it's done and
should not be scheduled again. And we only want to do that when
waiting for a reboot (if we have to wait).

`mender_client_check_deployment()` returns `MENDER_DONE` in case
there is no deployment so we need to translate that to
`MENDER_OK` signalling we are ok to proceed and check again next
time.

Ticket: MEN-7536
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>