fix: Move artifact_name cache to storage layer
On Artifact commit, the storage is updated with the new Artifact name
and the API code (both deployments/next and inventory) would then be
using an outdated value.

Modify the API code to get the value from the store, and move the cache
mechanism to the store itself.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>