chore: Use Zephyr logging macros directly
Zephyr logging is a complex thing that does a lot of magic and
optimizations. Using the macros from one place with a single '%s'
argument leaves no room for magic. By using the macros directly
from various places in our sources and with the various
arguments, we give the magic a chance to happen.

Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>