chore: Do not cache identity info
Identity of a device should not change. But since it is provided
by to us by a piece of code, we want to be able to detect issues
in that code. So if an update breaks that code, we want to reject
such an update and for that we need to call the code.

We are caching the authentication data (the JWT token) so the
identity callback is only called in case of re-authentication
which doesn't happen too often.

Ticket: MEN-7552
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>