fix: Detect potential reboot failure after failing client init
If the client initialization fails and needs a reboot, the
user-code-supplied callback is used to reboot the
system. However, if that callback fails to reboot, the client is
stuck. By switching the client into the `PENDING_REBOOT` state,
we have a chance to detect such hang and react to it.

Ticket: MEN-7811
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>