poc: Generate the Mender Artifact as a post-build action
Use existing hooks from Zephyr project to trigger the Artifact build as
a "Build Event".

CMake docs:
* https://cmake.org/cmake/help/latest/command/add_custom_command.html#build-events

This method does not really work because Mender needs the signed binary
and this one is in turn generated as a Build Event of the main binary.
CMake/Zephyr does not support having a build event depending of another
build event.

Ticket: MEN-8011