poc: Generate the Artifact as a "build always" target instead
See previous commit.

We can workaround the dependency limitation by having a build always
target to generate the Artifact, but do it only when the actual binary
has changed. And for this latter part, what we do is to hook into the
Build Event a command that just removes the potentially stalled
Artifact.

A bit hacky but gets the job done.

Ticket: MEN-8011