chore: Split the headers into public and internal ones
As part of preparation for defining our public API with the
public headers being trimmed down to only the things which need
to be part of the public API.

For the headers where we need both a public and a private one,
the strategy is to `#include` the public header from the private
one. For now, keeping everything in the public one and only
having the private one in place for easier/simpler `#include` in
the sources.

Private headers have `_PRIV_` in their include guards to avoid
conflicts with the public ones.

Ticket: MEN-7692
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>