ci: use `CI_COMMIT_SHA` as default value for `MENDER_MCU_REVISION`
If we manually trigger a pipeline on a PR branch,
`CI_EXTERNAL_PULL_REQUEST_SOURCE_BRANCH_SHA` will be empty,
resulting in the downstream pipeline to fail.

Furthermore, specifying which revision of mender-mcu to use, is done by
selecting which branch to run the pipeline for. Since we need the sha to
be able to checkout PR branches in mender-mcu-integration, we can simply
always set `MENDER_MCU_REVISION` to the sha of the latest commit.

Add a description to the variables you can modify to make it
easier when manually creating pipelines.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>