fix: Do early integrity checks for the non-payload items
We shall check `version` and `header.tar` as soon as they are fully
downloaded, so that the rest of the download can be aborted if the
integrity of any of these is compromised.

Implemented by splitting the integrity checks into individual items (to
be used during header parsing) and the remaining ones (to be used after
the payload(s) have been downloaded).

Changelog: None
Ticket: MEN-8075

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>