feat: Rework the inventory API to use callbacks
Instead of having a function to set the inventory data, we now
allow the user code to add callbacks the Mender client uses when
it needs inventory data. Such callbacks and the data they provide
can be marked as persistent, which means static data that doesn't
change and is thus only obtained and sent once, or dynamic, which
means fresh data is always obtained and sent at the specified
inventory refresh interval.

Ticket: MEN-7598
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>