fix: Cache the artifact name read from NVS
Even in case the artifact name is not known (e.g. the initial
state of a device until the first successful deployment) when the
string `"unknown"` is returned. We need to cache this value to
fix two issues:

1. Make it cached and avoid trying to read it (and potentially
   fail) again and again and
2. avoid memory leaks when the callers of the
   `mender_storage_get_artifact_name()` function get a `const`
   pointer and don't `free()` it

We then don't need (cannot, in fact) `free()` the value in
`api_check_for_deployment_v2()`.

Ticket: MEN-8056
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>