chore: Merge the `flash` code into the zephyr-image update module
The POSIX implementation is just an artifical one that has no
real usage (and the *weak* one as well, of course). The only real
flashing implementation is the Zephyr/MCUBoot one. By only
keeping the latter, we can get rid of the unnecessary abstraction
and sources and simplify things.

Ticket: MEN-7554
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>