fix: erase image in slot 1 on failed deployment
Once a new boot partition has been requested, the swap method is set to
`test`. If we fail a deployment in e.g. `MENDER_UPDATE_STATE_INSTALL`,
the swap method remains, which means that on a reboot after a
failed deployment, the device will attempt to boot into the image in
slot 1. We'll therefore need to erase the image in slot 1 if we fail the
deployment and we're in a confirmed image.

Ticket: MEN-8215
Changelog: Title

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>