feat: Generate the Mender Artifact as part of the Zephyr build byproducts
Integrate with CMake build system so that a Mender Artifact is generated
next to the Zephyr kernel and its byproducts.

For the most simple use case, only the Artifact Name needs to be
specified, the  rest are optional or have sane default values. This
policy imitates the design of other Mender build integrations, like
Yocto or mender-convert.

Ticket: MEN-8102
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>