chore: Make inventory "feature" hard to disable
While we want to keep the code around for potentially minimalistic
builds of the Client, we at Mender team consider it a core feature that
should not be opted out.

The middle point is to mark the disabling of the inventory as an
EXPERIMENTAL feature (we won't be testing it, after all) but keep the
code around for hackers or advance users to play with.

Ticket: MEN-7548

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>