feat: Allow using a different storage partition than `storage_partition`
The best solution, that we should encourage, is to use a separate
`mender_partition` for Mender non-volatile data. However, that
requires modifying device tree so we also need to allow using the
`storage_partition` partition that is available in most boards'
device trees in Zephyr. And if the application code wants to use
a completely different (custom) partition, we should allow that
too.

In all cases, we should allow the application code to say that it
wants Mender data to start at a specific sector in the partition
so that it can use the first N sectors for other purposes.

Ticket: MEN-8243
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>