fix: store `MENDER_UPDATE_STATE_REBOOT` before calling install callback
In the `zephyr-image` update module we set the pending image in
the install callback. If we don't store the next state before
setting the pending image, a spontaneous reboot in
`MENDER_UPDATE_STATE_INSTALL` will simply go to failure and stay in the
unconfirmed image in slot 1. By storing the next state, we assure that
the deployment will be marked as a failure and that we don't end up in
the wrong image.

Changelog: None
Ticket: None

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>