fix: Ensure aborted deployment is never committed
An aborted deployment should never be committed. However, the
only way to check if a deployment was aborted or not is to try to
update its status and check the response status code. To leverage
that, we set the deployment status to `Installing` once again
right before status. And if that update tells us the deployment
was aborted, we don't commit it.

Ticket: MEN-8223
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>