fix: reschedule work items with backoff on retry errors
If `MENDER_RETRY_ERROR` is detected in a work function it will be retried
with a separate backoff interval which doubles on each try. It will
continue until it reaches the max backoff interval and keep
retrying using the max backoff interval.

Added configurable options for backoff:
* `MENDER_RETRY_ERROR_BACKOFF`
* `MENDER_RETRY_ERROR_MAX_BACKOFF`

Ticket: MEN-8222
Changelog: Title

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>